[ISSUE-19] Refactoring TransformerBuilder implementations to have the same base.
